feat: CRUD destinos

+++ src/main/java/com/challenge/jornadamilhasapi/dtos/destino/DadosCadastroDestinoDTO.java
@@ -3,5 +3,7 @@ package com.challenge.jornadamilhasapi.dtos.destino;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 
-public record DadosCadastroDestinoDTO(@NotBlank String foto, @NotBlank String nome, @NotNull Double preco) {
+import java.math.BigDecimal;
+
+public record DadosCadastroDestinoDTO(@NotBlank String foto, @NotBlank String nome, @NotNull BigDecimal preco) {
 }
